+++ spec/support/custom_object_changes_adapter.rb
@@ -10,6 +10,10 @@ class CustomObjectChangesAdapter
     version.changeset
   end
 
+  def where_attribute_changes(klass, attribute)
+    klass.where(attribute)
+  end
+
   def where_object_changes(klass, attributes)
     klass.where(attributes)
   end